+++ libtc/__init__.py
@@ -1,11 +1,12 @@
-from .bencode import bdecode, bencode
+from .bencode import BTFailure, bdecode, bencode
 from .clients import *
 from .exceptions import FailedToExecuteException, LibTorrentClientException
 from .management import move_torrent
+from .parse_clients import parse_clients_from_toml_dict
 from .torrent import *
 from .utils import TorrentProblems
 
-__version__ = "1.2.1"
+__version__ = "1.2.2"
 
 __all__ = [
     "DelugeClient",
@@ -25,4 +26,6 @@ __all__ = [
     "move_torrent",
     "parse_libtc_url",
     "TorrentProblems",
+    "parse_clients_from_toml_dict",
+    "BTFailure",
 ]